food_delivery/requirements.txt
Django==4.2.7
pymongo==4.5.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.1
python-dotenv==1.0.0
Pillow==10.0.1
